// src/types/IProfile.ts
export default interface IProfile {
  
  uuid?: string;
  student_id?: number;
  firstname?: string;
  lastname?: string;
  description?: string;
  place_of_birth?: string;
  skill?: string;
  avatar?: string;
  address?: string;
  city_code?: string;
  date_of_birth?: string;
  district_code?: string;
  facebook?: string;
  instagram?: string;
  linkedin?: string;
  twitter?: string;
  gender?: string;
  house_phone?: string;
  mobile_phone?: string;
  work_phone?: string;
  nationality?: string;
  postal_code?: string;
  province_code?: string;
  village_code?: string;
  email?: string;

}